+++ NEWS.md
@@ -16,7 +16,7 @@
 * added new `speciesPresentFromNTEMS` function to import dominant species layer from NTEMS and create factor raster to be hosted on google drive
 * added new `prepInputs_NTEMS_DominantSpecies` function for importing dominant species layers from NTEMS website
 * updated `speciesInStudyArea` function to create `dataSource` parameter to direct function to download KNN or NTEMS factor raster from google drive and create associated species list  
-* updated `prepRawBiomassMap` function to allow for incorporation of SCANFI biomass
+* updated `prepRawBiomassMap` function to allow for incorporation of NTEMS or SCANFI biomass
 
 # LandR 1.1.5